Amend the negative greater than positive votes clause in "Interpreting the Vote"
There may be a situation where the following happens:

 * An election for a candidate has started, with 0 total votes.
 * An opposing voter casts their votes against the candidate, leaving the poll at
   0 positive votes and 1 negative vote.
 * Due to this clause, the election fails immediately because there are a greater
   amount of negative votes than positive votes (1 negative > 0 positive).

This allows (even accidentally) for a voter to singlehandedly fail an election, even
if the candidate would have been passed by the whole voter pool.

To prevent this possibility, the condition is amended to fail the election if an
absolute majority of the voters have voted against the candidate, instead of only a
simple majority.