RFC 7: Update voting thresolds
Our current voting system has various flaws that I detailed in a staff chat, copied below

> Using our current count of eligible voters of 14, if a pool is 10 👍 0 👎 3 🤷, and the last voter is against, they would rather put a 🤷 to reach the abstaining threshold than a 👎 which would have a very tiny effect. Arguably that's an extreme case, but with one more voter, it is the same for 8 👍 3 👎 3 🤷. 
> 
> It is also worth noting that 9 👍 5 👎 0 🤷 is considered a pass, which already seems quite controversial, while 9 👍 1 👎 4 🤷 is considered a fail.
> 
> One last, perhaps less note-worthy flaw is we don't account for people simply not voting. This means that 4 👍 0 👎 0 🤷 is considered a pass, while it is clearly not enough people gave their thoughts.

This suggestion uses numerical thresholds that aim to simplify the voting process, along with making it more resilient and hopefully fairer.

It also removes the need for admins as tie breakers. I decided to remove that paragraph entirely from the RFC, which may be an unpopular change.