Kill community centrict RFCs
While Quilt was initially trying to preserve its structure through RFCs, we've discovered this isn't an effective way to achieve this goal. Those RFCs have historically always been inaccurate compared to what's in action, because of the inability to move and iterate quickly due to all the legal-ish wording, procedures, and RFCs.

We've decided to give ourselves back the ability to be more agile with our structure, without loosing the core concepts that make up Quilt.